Refactor psllq_r128

+++ src/native/sse_instr.c
@@ -224,9 +224,9 @@ void psllq_r128(int32_t r, uint32_t shift)
 
     if(shift <= 31) {
         result.u32[0] = destination.u32[0] << shift;
-        result.u32[1] = destination.u32[1] << shift | (((uint32_t) destination.u32[0]) >> (32 - shift));
+        result.u32[1] = destination.u32[1] << shift | (destination.u32[0] >> (32 - shift));
         result.u32[2] = destination.u32[2] << shift;
-        result.u32[3] = destination.u32[3] << shift | (((uint32_t) destination.u32[2]) >> (32 - shift));
+        result.u32[3] = destination.u32[3] << shift | (destination.u32[2] >> (32 - shift));
     }
     else if(shift <= 63) {
         result.u32[0] = 0;
